pyproject.toml
[tool.poetry]
name = "ai-fitness-coach"
version = "0.1.0"
description = "An AI-powered personal fitness coach that integrates with your calendar and health data to generate adaptive workout plans."
authors = ["Sophia Malnight-Alvarez <smalnightalvarez@gmail.com>"]
readme = "README.md"
packages = [{include = "fitness_coach", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "^0.3.25"
langchain-openai = "^0.3.19"
langchain-community = "^0.3.24"
langgraph = "^0.4.8"
streamlit = "^1.45.1"
openai = "^1.84.0"
ollama = "^0.5.1"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.11.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

